feat: Room에 Reservation 데이터 저장 및 불러오기

+++ app/src/main/java/woowacourse/movie/presentation/ui/seatselection/SeatSelectionContract.kt
@@ -8,7 +8,7 @@ interface SeatSelectionContract {
     interface View : BaseView, SeatSelectionActionHandler {
         fun showSeatModel(seatModel: SeatSelectionUiModel)
 
-        fun navigateToReservation(reservationId: Int)
+        fun navigateToReservation(reservationId: Long)
 
         fun showReservationDialog()
